+++ packages/tokens/themes/index.ts
@@ -1,12 +1,10 @@
 import { ignite } from "./ignite";
-import { test } from "./themeTest";
 import { memberflix } from "./memberflix";
 
 export * from "./ignite";
-export * from "./themeTest";
+export * from "./memberflix";
 
 export default {
   ignite,
-  test,
   memberflix,
 };