+++ packages/tokens/themes/themeTest/index.ts
@@ -1,8 +1,0 @@
-import { testColors } from "./testColors";
-import { testFonts } from "./testFonts";
-
-export const test = {
-  title: "Test",
-  colors: testColors,
-  fonts: testFonts,
-};