+++ packages/tokens/themes/themeTest/testColors.ts
@@ -1,6 +1,0 @@
-export const testColors = {
-  test300: "#07E",
-  test500: "#05F",
-  test700: "#043",
-  test900: "#01D",
-};