+++ packages/tokens/themes/themeTest/testFonts.ts
@@ -1,4 +1,0 @@
-export const testFonts = {
-  default: "Helvetica, sans-serif",
-  batata: "monospace",
-};