+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rbeiro-ui/tokens
 
+## 2.0.0
+
+### Major Changes
+
+- Fixed the export on /tokens
+
+### Minor Changes
+
+- 4eab387: Corrected Memberflix theme variable names
+
 ## 1.2.0
 
 ### Minor Changes